first session of code cleaning
